alter: price data type

+++ cafe/models.py
@@ -14,7 +14,7 @@ class MenuItem(models.Model):
     name = models.CharField(max_length=100)
     description = models.TextField(blank=True)
     image = models.ImageField(upload_to="menu_images/", blank=True, null=True)
-    price = models.DecimalField(max_digits=8, decimal_places=2)
+    price = models.IntegerField()
     stock = models.PositiveIntegerField(default=0)
 
     def __str__(self):
